+++ Src/CSharpier.Rider/gradle.properties
@@ -4,7 +4,7 @@
 pluginGroup = com.intellij.csharpier
 pluginName = csharpier
 # SemVer format -> https://semver.org
-pluginVersion = 1.2.0-beta4
+pluginVersion = 1.2.0-beta5
 
 # See https://plugins.jetbrains.com/docs/intellij/build-number-ranges.html
 # for insight into build numbers and IntelliJ Platform versions.
@@ -12,8 +12,8 @@ pluginSinceBuild = 203
 pluginUntilBuild = 213.*
 
 # IntelliJ Platform Properties -> https://github.com/JetBrains/gradle-intellij-plugin#intellij-platform-properties
-platformType = IC
-platformVersion = 2020.3.4
+platformType = RD
+platformVersion = 2021.2.1
 
 # Plugin Dependencies -> https://plugins.jetbrains.com/docs/intellij/plugin-dependencies.html
 # Example: platformPlugins = com.intellij.java, com.jetbrains.php:203.4449.22
