+++ Src/CSharpier.Rider/src/main/resources/META-INF/plugin.xml
@@ -7,26 +7,30 @@
     <depends>com.intellij.modules.rider</depends>
 
     <extensions defaultExtensionNs="com.intellij">
+        <applicationService serviceImplementation="com.intellij.csharpier.ReformatWithCSharpierOnSave"/>
         <projectService serviceImplementation="com.intellij.csharpier.CSharpierProcessProvider"/>
         <projectService serviceImplementation="com.intellij.csharpier.FormattingService"/>
         <projectService serviceImplementation="com.intellij.csharpier.CSharpierSettings"/>
         <projectService serviceImplementation="com.intellij.csharpier.InstallerService"/>
-<!--        <projectService serviceImplementation="com.intellij.csharpier.ReformatWithCSharpierOnSave4"/>-->
         <notificationGroup id="CSharpier" displayType="STICKY_BALLOON"/>
         <postStartupActivity implementation="com.intellij.csharpier.CSharpierStartup"/>
-        <fileDocumentManagerListener implementation="com.intellij.csharpier.ReformatWithCSharpierOnSave3"/>
-<!--        <actionOnSave id="ReformatWithCSharpierOnSave2"-->
-<!--                      implementation="com.intellij.csharpier.ReformatWithCSharpierOnSave2"-->
-<!--                      order="after FormatOnSaveAction"/>-->
+<!--        <fileDocumentManagerListener implementation="com.intellij.csharpier.ReformatWithCSharpierOnSave3"/>-->
         <projectConfigurable
                 parentId="tools"
                 id="com.intellij.csharpier.CSharpierSettingsComponent"
                 instance="com.intellij.csharpier.CSharpierSettingsComponent"/>
     </extensions>
     <projectListeners>
-<!--        <listener class="com.intellij.csharpier.ReformatWithCSharpierOnSave"-->
+<!--        <listener class="com.intellij.csharpier.ReformatWithCSharpierOnSaveOld"-->
 <!--                  topic="com.intellij.openapi.fileEditor.FileDocumentManagerListener"/>-->
     </projectListeners>
+
+<!--    <application-components>-->
+<!--        <component>-->
+<!--            <implementation-class>com.intellij.csharpier.ReformatWithCSharpierOnSave</implementation-class>-->
+<!--        </component>-->
+
+<!--    </application-components>-->
 
     <actions>
         <action id="com.intellij.csharpier.ReformatWithCSharpierAction"
